add task

+++ src/controllers/notesController.js
@@ -5,7 +5,9 @@ import createHttpError from "http-errors";
 export const getAllNotes = async (req, res) => {
     const {tag, search, page = 1, perPage = 10} = req.query;
 
-    const notesQuery = Note.find();
+    const notesQuery = Note.find({
+        userId: req.user._id,
+    });
 
     const skip = (page - 1) * perPage;
 
@@ -39,7 +41,10 @@ export const getAllNotes = async (req, res) => {
 
 export const getNoteById = async (req, res) => {
     const {noteId} = req.params;
-    const note = await Note.findById(noteId);
+    const note = await Note.findById({
+        noteId,
+        userId: req.user._id,
+    });
 
     if (!note) {
         throw createHttpError(404, 'Note not found');
@@ -50,13 +55,16 @@ export const getNoteById = async (req, res) => {
 
 
 export const createNote = async (req, res) => {
-    const note = await Note.create(req.body);
+    const note = await Note.create({
+        ...req.body,
+        userId: req.user._id
+    });
     res.status(201).json(note);
 };
 
 export const deleteNote = async (req, res) => {
     const {noteId} = req.params;
-    const note = await Note.findOneAndDelete({_id: noteId});
+    const note = await Note.findOneAndDelete({_id: noteId, userId: req.user._id});
     if (!note) {
         throw createHttpError(404, "Note not found");
     }
@@ -65,7 +73,7 @@ export const deleteNote = async (req, res) => {
 
 export const updateNote = async (req, res) => {
     const {noteId} = req.params;
-    const note = await Note.findOneAndUpdate({_id: noteId}, req.body, {returnDocument: "after"});
+    const note = await Note.findOneAndUpdate({_id: noteId, userId: req.user._id}, req.body, {returnDocument: "after"});
     if (!note) {
         throw createHttpError(404, "Note not found");
     }
